+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-doc
 
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.9
+
 ## 0.6.7
 
 ### Patch Changes