+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-export
 
+## 0.6.9
+
+### Patch Changes
+
+- add bytecode and linkedData to export
+- Updated dependencies
+  - rocketh@0.6.9
+
 ## 0.6.8
 
 ### Patch Changes